requirements.txt
# BROME
wheel
Pillow
tornado
netifaces
selenium
decorator
pudb==2014.1
PyYAML==3.11
ipython==3.0.0
lxml==3.4.1
cssselect==0.9.1
psutil==2.1.3
paramiko==1.15.1
boto==2.32.1
pyvbox==0.2.0

# BROME-WEBSERVER
aiohttp==0.21.6
python-memcached==1.58
requests
python-dateutil==2.5.3
cchardet==1.0.0
aiohttp-debugtoolbar==0.1.1
aiohttp-session==0.5.0
aiohttp-jinja2==0.7.0
aioredis==0.2.7

## MAC
## env LDFLAGS="-L$(brew --prefix openssl)/lib" CFLAGS="-I$(brew --prefix openssl)/include" pip install cryptography
cryptography

commentjson
MongoAlchemy
bcrypt
validate_email
prometheus_client
rq
redis
jinja2
Flask
Flask-Admin
Flask-Login>=0.3.0
flask-appbuilder
flask-mongoengine
Flask-Session
